Get rid of Motorized in favor of findMotor()
