+++ components/kernel/drivers/MotorDriver.cpp
@@ -1,0 +1,11 @@
+#include "MotorDriver.hpp"
+
+namespace farmhub::kernel::drivers {
+
+MotorPhase operator-(MotorPhase phase) {
+    return phase == MotorPhase::FORWARD
+        ? MotorPhase::REVERSE
+        : MotorPhase::FORWARD;
+}
+
+}
